Add fluent validation for proper http model validation

+++ ShareJobsData/src/ShareJobsDataCli/GitHub/Artifacts/CurrentWorkflowRun/HttpModels/Requests/GitHubFinalizeArtifactContainerRequest.cs
@@ -1,11 +1,13 @@
+using ShareJobsDataCli.ArgumentValidations;
+
 namespace ShareJobsDataCli.GitHub.Artifacts.CurrentWorkflowRun.HttpModels.Requests;
 
 internal sealed record GitHubFinalizeArtifactContainerRequest
 {
-    public GitHubFinalizeArtifactContainerRequest(int artifactContainerSize)
+    public GitHubFinalizeArtifactContainerRequest(long artifactContainerSize)
     {
         Size = artifactContainerSize.Positive();
     }
 
-    public int Size { get; }
+    public long Size { get; }
 }
